wip: rewrite piano
Now Resources and Options are package-wise

+++ scene/options.go
@@ -57,7 +57,7 @@ type ScreenOptions struct {
 	// whereas ScreenSize is the logical size of the screen.
 	Resolution           draws.XY
 	Fullscreen           bool
-	BackgroundBrightness float64
+	BackgroundBrightness float32
 	DebugPrint           bool
 }
